BLIF: Add dummy circuit component & refactor 'generic port interface' code from the VHDL simulator
