+++ src/main/java/com/cburch/logisim/std/hdl/VhdlEntityComponent.java
@@ -15,21 +15,15 @@ import com.cburch.hdl.HdlModel;
 import com.cburch.hdl.HdlModelListener;
 import com.cburch.logisim.data.Attribute;
 import com.cburch.logisim.data.AttributeSet;
-import com.cburch.logisim.data.Bounds;
 import com.cburch.logisim.data.Value;
 import com.cburch.logisim.gui.icons.ArithmeticIcon;
 import com.cburch.logisim.gui.main.Frame;
 import com.cburch.logisim.instance.Instance;
-import com.cburch.logisim.instance.InstanceFactory;
-import com.cburch.logisim.instance.InstancePainter;
 import com.cburch.logisim.instance.InstanceState;
+import com.cburch.logisim.instance.Port;
 import com.cburch.logisim.instance.StdAttr;
-import com.cburch.logisim.prefs.AppPreferences;
-import com.cburch.logisim.util.GraphicsUtil;
-import com.cburch.logisim.util.StringUtil;
 import com.cburch.logisim.vhdl.base.VhdlSimConstants;
 import com.cburch.logisim.vhdl.sim.VhdlSimulatorTop;
-import java.awt.Color;
 import java.awt.Window;
 import java.io.IOException;
 import java.io.PrintWriter;
@@ -38,7 +32,7 @@ import java.util.WeakHashMap;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class VhdlEntityComponent extends InstanceFactory {
+public class VhdlEntityComponent extends GenericInterfaceComponent {
   /**
    * Unique identifier of the tool, used as reference in project files.
    * Do NOT change as it will prevent project files from loading.
@@ -97,11 +91,6 @@ public class VhdlEntityComponent extends InstanceFactory {
   static final Logger logger = LoggerFactory.getLogger(VhdlEntityComponent.class);
 
   public static final Attribute<VhdlContentComponent> CONTENT_ATTR = new ContentAttribute();
-  static final int WIDTH = 140;
-  static final int HEIGHT = 40;
-  static final int PORT_GAP = 10;
-
-  static final int X_PADDING = 5;
 
   private final WeakHashMap<Instance, VhdlEntityListener> contentListeners;
 
@@ -159,76 +148,28 @@ public class VhdlEntityComponent extends InstanceFactory {
   }
 
   @Override
-  public Bounds getOffsetBounds(AttributeSet attrs) {
+  protected String getGIAttributesName(AttributeSet attrs) {
     final var content = attrs.getValue(CONTENT_ATTR);
-    final var nbInputs = content.getInputsNumber();
-    final var nbOutputs = content.getOutputsNumber();
+    return content.getName();
+  }
 
-    return Bounds.create(0, 0, WIDTH, Math.max(nbInputs, nbOutputs) * PORT_GAP + HEIGHT);
+  @Override
+  protected Port[] getGIAttributesInputs(AttributeSet attrs) {
+    final var content = attrs.getValue(CONTENT_ATTR);
+    return content.getInputs();
+  }
+
+  @Override
+  protected Port[] getGIAttributesOutputs(AttributeSet attrs) {
+    final var content = attrs.getValue(CONTENT_ATTR);
+    return content.getOutputs();
   }
 
   @Override
   protected void instanceAttributeChanged(Instance instance, Attribute<?> attr) {
     if (attr == CONTENT_ATTR) {
       updatePorts(instance);
-      instance.recomputeBounds();
     }
-  }
-
-  @Override
-  public void paintInstance(InstancePainter painter) {
-    final var g = painter.getGraphics();
-    final var content = painter.getAttributeValue(CONTENT_ATTR);
-    var metric = g.getFontMetrics();
-
-    final var bds = painter.getBounds();
-    final var x0 = bds.getX() + (bds.getWidth() / 2);
-    final var y0 = bds.getY() + metric.getHeight() + 12;
-    g.setColor(new Color(AppPreferences.COMPONENT_COLOR.get()));
-    GraphicsUtil.drawText(
-        g,
-        StringUtil.resizeString(content.getName(), metric, WIDTH),
-        x0,
-        y0,
-        GraphicsUtil.H_CENTER,
-        GraphicsUtil.V_BOTTOM);
-
-    final var glbLabel = painter.getAttributeValue(StdAttr.LABEL);
-    if (glbLabel != null) {
-      final var font = g.getFont();
-      g.setFont(painter.getAttributeValue(StdAttr.LABEL_FONT));
-      GraphicsUtil.drawCenteredText(
-          g, glbLabel, bds.getX() + bds.getWidth() / 2, bds.getY() - g.getFont().getSize());
-      g.setFont(font);
-    }
-
-    g.setColor(new Color(AppPreferences.COMPONENT_SECONDARY_COLOR.get()));
-    g.setFont(g.getFont().deriveFont((float) 10));
-    metric = g.getFontMetrics();
-
-    final var inputs = content.getInputs();
-    final var outputs = content.getOutputs();
-
-    for (var i = 0; i < inputs.length; i++)
-      GraphicsUtil.drawText(
-          g,
-          StringUtil.resizeString(inputs[i].getToolTip(), metric, (WIDTH / 2) - X_PADDING),
-          bds.getX() + 5,
-          bds.getY() + HEIGHT - 2 + (i * PORT_GAP),
-          GraphicsUtil.H_LEFT,
-          GraphicsUtil.V_CENTER);
-    for (var i = 0; i < outputs.length; i++)
-      GraphicsUtil.drawText(
-          g,
-          StringUtil.resizeString(outputs[i].getToolTip(), metric, (WIDTH / 2) - X_PADDING),
-          bds.getX() + WIDTH - 5,
-          bds.getY() + HEIGHT - 2 + (i * PORT_GAP),
-          GraphicsUtil.H_RIGHT,
-          GraphicsUtil.V_CENTER);
-
-    g.setColor(new Color(AppPreferences.COMPONENT_COLOR.get()));
-    painter.drawBounds();
-    painter.drawPorts();
   }
 
   /**
@@ -342,10 +283,5 @@ public class VhdlEntityComponent extends InstanceFactory {
       logger.error("Could not create vhdl file: {}", e.getMessage());
       e.printStackTrace();
     }
-  }
-
-  void updatePorts(Instance instance) {
-    VhdlContentComponent content = instance.getAttributeValue(CONTENT_ATTR);
-    instance.setPorts(content.getPorts());
   }
 }
