+++ src/main/resources/resources/logisim/strings/hdl/hdl.properties
@@ -28,6 +28,10 @@ validateAndSaveButton = Validate & Save
 #
 hdlSaveDialog = Export VHDL file
 #
+# hdl/BlifCircuitComponent.java
+#
+blifComponent = BLIF Component
+#
 # hdl/Hdl.java
 #
 hdlLibrary = HDL-IP
